refactor: Accomodate course requirement of 15 days
Previous code could only calculate till 6 days of learning. There
were 6 days worth of temp tables each. Hackerrank required
calculation of 15 days of data. 15 temp tables would have to be
added to accomodate that, plus the code wouldn't be dynamic if
more days worth of data were added.

With the new temp table, it will calculate as many days worth of
data is included, provided it is for the same month.

If the table includes data for different months, the code will not
work. This program only accomodates 15 days of learning, as stated
in the problem.